transfer to notebook
transfer, no time to explain.

+++ Assets/Scripts/SciptableObjects/DataSO.cs
@@ -8,5 +8,5 @@ public class DataSO : ScriptableObject
     public int age;
 
     public string[] responses;
-    public int[] frequencies;
+    public double[] frequencies;
 }
